server/requirements.txt
click==8.1.2
Flask==2.1.1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Werkzeug==2.1.1
zipp==3.8.0
google-auth-oauthlib==0.4.6
Flask-SQLAlchemy==2.5.1
loguru==0.6.0
cryptography==36.0.0
Flask-Migrate==3.1.0
google-api-python-client==2.23.0
gphotospy==0.2.1